Backend/requirements.txt
flask
pytesseract
pillow
pymupdf
sentence-transformers
faiss-cpu
requests
